+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added
   - Commands for copying hub/project/folder/item/version IDs to clipboard
+  - Support for additional webhooks
 - Fixed
   - Object upload using the new "direct S3" API